refactor: create src/cli/admin module, migrate handlers from handlers.rs
Create src/cli/admin/ directory with submodules (agent, config, rule,
skill, stop, common) to organize admin handler functions by feature.
Remove src/handlers.rs. Update src/main.rs to import from
closeclaw::cli::admin::* instead of crate::handlers. Gate test module
with #[cfg(test)].

Source: user
Type: refactor
