+++ src/cli/mod.rs
@@ -1,5 +1,6 @@
 //! CLI Tool - closeclaw command-line interface
 
+pub mod admin;
 pub mod args;
 pub mod chat;
 pub mod config_wizard;
